Extract the event loop functionality from WeakRealmNotifier
This makes it possible to reuse it for other things and cuts down on the
duplicated code a bit.
