+++ src/impl/weak_realm_notifier.hpp
@@ -1,6 +1,6 @@
 ////////////////////////////////////////////////////////////////////////////
 //
-// Copyright 2016 Realm Inc.
+// Copyright 2015 Realm Inc.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -19,16 +19,57 @@
 #ifndef REALM_WEAK_REALM_NOTIFIER_HPP
 #define REALM_WEAK_REALM_NOTIFIER_HPP
 
-#include <realm/util/features.h>
+#include <memory>
+#include <thread>
 
-#if REALM_PLATFORM_NODE
-#include "impl/node/weak_realm_notifier.hpp"
-#elif REALM_PLATFORM_APPLE
-#include "impl/apple/weak_realm_notifier.hpp"
-#elif REALM_ANDROID
-#include "impl/android/weak_realm_notifier.hpp"
-#else
-#include "impl/generic/weak_realm_notifier.hpp"
-#endif
+namespace realm {
+class Realm;
+
+namespace util {
+template<typename> class EventLoopSignal;
+}
+
+namespace _impl {
+// WeakRealmNotifier stores a weak reference to a Realm instance, along with all of
+// the information about a Realm that needs to be accessed from other threads.
+// This is needed to avoid forming strong references to the Realm instances on
+// other threads, which can produce deadlocks when the last strong reference to
+// a Realm instance is released from within a function holding the cache lock.
+class WeakRealmNotifier {
+public:
+    WeakRealmNotifier(const std::shared_ptr<Realm>& realm, bool cache);
+    ~WeakRealmNotifier();
+
+    // Get a strong reference to the cached realm
+    std::shared_ptr<Realm> realm() const { return m_realm.lock(); }
+
+    // Does this WeakRealmNotifier store a Realm instance that should be used on the current thread?
+    bool is_cached_for_current_thread() const { return m_cache && is_for_current_thread(); }
+
+    // Has the Realm instance been destroyed?
+    bool expired() const { return m_realm.expired(); }
+
+    // Is this a WeakRealmNotifier for the given Realm instance?
+    bool is_for_realm(Realm* realm) const { return realm == m_realm_key; }
+
+    bool is_for_current_thread() const { return m_thread_id == std::this_thread::get_id(); }
+
+    void notify();
+
+private:
+    std::weak_ptr<Realm> m_realm;
+    std::thread::id m_thread_id = std::this_thread::get_id();
+    void* m_realm_key;
+    bool m_cache = false;
+
+    struct Callback {
+        std::weak_ptr<Realm> weak_realm;
+        void operator()();
+    };
+    std::shared_ptr<util::EventLoopSignal<Callback>> m_signal;
+};
+
+} // namespace _impl
+} // namespace realm
 
 #endif // REALM_WEAK_REALM_NOTIFIER_HPP
